Separate main into Examples
